requirements.txt
openai-whisper
torch
tqdm
numpy
regex
requests